[jit] improve error message on inferred type
Previously we would give confusing error messages when we inferred an
unannotated type to be Tensor. This adds a hint to the compiler
diagnostic so that users might be less confused.

This PR moves the type inference logic from the frontend to
compiler.cpp. So now the `Param` tree view's type is optional, and we
track whether or not we inferred the type of a schema argument to be
Tensor. We use that info to format a proper error message.

Also this dedupes some of the error messages reporting since we were
replicating similar strings in several places.

gh-metadata: pytorch pytorch 21058 gh/suo/42/head